requirements.txt
django==3.2
djangorestframework==3.12.0
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
flake8-commas==2.1.0
flake8-import-order==0.18.1
black==22.3.0